src/etl/requirements.txt
elasticsearch==7.17.9
pydantic==2.3.0 
pydantic-settings==2.0.3 
python-dotenv==1.0
redis==4.4.4
psycopg2-binary==2.9.6